requirements.txt
fastapi~=0.120.0
uvicorn[standard]
python-multipart
sqlalchemy~=2.0.44
langchain
langchain_google_genai
python-dotenv~=1.1.1
unstructured
python-docx
pypdf
langchain_community
psycopg2-binary
langchain_core

pydantic~=2.12.3
langchain-google-genai~=2.0.10
langchain-core~=0.3.79
langchain-community~=0.3.31
protobuf~=5.29.5
faiss-cpu
